+++ R/Spectra-functions.R
@@ -13,10 +13,9 @@ NULL
 #'
 #' @importClassesFrom ProtGenerics ProcessingStep
 #'
-#' @rdname Spectra
+#' @importFrom methods slotNames
 #'
-#' UPDATE: with `spectraVariables` we allow the user to pass the names of the
-#' spectraVariables he/she wishes to pass along to the function.
+#' @rdname Spectra
 addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
     if (missing(FUN))
         return(object)
@@ -29,6 +28,19 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
     validObject(object)
     object
 }
+
+## .apply_processing_queue <- function(x, msLevel, centroided, queue = NULL) {
+##     if (length(queue)) {
+##         for (i in seq_along(x)) {
+##             for (pStep in queue) {
+##                 x[[i]] <- executeProcessingStep(pStep, x[[i]],
+##                                                 spectrumMsLevel = msLevel[i],
+##                                                 centroided = centroided[i])
+##             }
+##         }
+##     }
+##     x
+## }
 
 #' @description
 #'
@@ -45,6 +57,9 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
 #'
 #' @param queue `list` of `ProcessStep` elements.
 #'
+#' @param spectraData `data.frame` with additional spectra variables as
+#'     requested with `addProcessing`.
+#'
 #' @return `list` of peak `matrix` elements.
 #'
 #' @author Johannes Rainer
@@ -52,24 +67,7 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
 #' @importFrom ProtGenerics executeProcessingStep
 #'
 #' @noRd
-.apply_processing_queue <- function(x, msLevel, centroided, queue = NULL) {
-    if (length(queue)) {
-        for (i in seq_along(x)) {
-            for (pStep in queue) {
-                x[[i]] <- executeProcessingStep(pStep, x[[i]],
-                                                spectrumMsLevel = msLevel[i],
-                                                centroided = centroided[i])
-            }
-        }
-    }
-    x
-}
-
-#' @param spectraData `data.frame` with additional spectra variables as
-#'     requested with `addProcessing`.
-#'
-#' @noRd
-.apply_processing_queue2 <- function(x, spectraData = data.frame(),
+.apply_processing_queue <- function(x, spectraData = data.frame(),
                                      queue = NULL) {
     if (length(queue)) {
         args <- list()
@@ -88,6 +86,33 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
     x
 }
 
+## .peaksapply <- function(object, FUN = NULL, ..., f = dataStorage(object),
+##                         BPPARAM = bpparam()) {
+##     len <- length(object)
+##     if (!len)
+##         return(list())
+##     if (length(f) != len)
+##         stop("Length of 'f' has to match 'length(object)' (", len, ")")
+##     if (!is.factor(f))
+##         f <- factor(f)
+##     pqueue <- object@processingQueue
+##     if (!is.null(FUN))
+##         pqueue <- c(pqueue, ProcessingStep(FUN, ARGS = list(...)))
+##     if (length(levels(f)) > 1 || length(pqueue)) {
+##         res <- bplapply(split(object@backend, f), function(z, queue) {
+##             metad <- spectraData(z, columns = c("msLevel", "centroided"))
+##             .apply_processing_queue(peaksData(z), metad$msLevel,
+##                                     metad$centroided, queue = queue)
+##         }, queue = pqueue, BPPARAM = BPPARAM)
+##         unsplit(res, f = f, drop = TRUE)
+##     } else peaksData(object@backend)
+## }
+
+.processingQueueVariables <- function(x) {
+    if (any(slotNames(x) == "processingQueueVariables"))
+        x@processingQueueVariables
+    else c("msLevel", "centroided")
+}
 
 #' @title Apply arbitrary functions and processing queue to peaks matrices
 #'
@@ -101,6 +126,9 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
 #' @param FUN optional function to be applied to the peaks matrix. The peaks
 #'     matrix will be passed to the first argument of the function. The function
 #'     should also have arguments `...`.
+#'
+#' @param spectraVariables `character` defining spectra variables that should
+#'     be passed to the function.
 #'
 #' @param ... optional additional arguments to `FUN`.
 #'
@@ -116,43 +144,9 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
 #' @importMethodsFrom BiocParallel bplapply
 #'
 #' @noRd
-.peaksapply <- function(object, FUN = NULL, ..., f = dataStorage(object),
-                        BPPARAM = bpparam()) {
-    len <- length(object)
-    if (!len)
-        return(list())
-    if (length(f) != len)
-        stop("Length of 'f' has to match 'length(object)' (", len, ")")
-    if (!is.factor(f))
-        f <- factor(f)
-    pqueue <- object@processingQueue
-    if (!is.null(FUN))
-        pqueue <- c(pqueue, ProcessingStep(FUN, ARGS = list(...)))
-    if (length(levels(f)) > 1 || length(pqueue)) {
-        res <- bplapply(split(object@backend, f), function(z, queue) {
-            metad <- spectraData(z, columns = c("msLevel", "centroided"))
-            .apply_processing_queue(peaksData(z), metad$msLevel,
-                                    metad$centroided, queue = queue)
-        }, queue = pqueue, BPPARAM = BPPARAM)
-        unsplit(res, f = f, drop = TRUE)
-    } else peaksData(object@backend)
-}
-
-.processingQueueVariables <- function(x) {
-    if (any(slotNames(x) == "processingQueueVariables"))
-        x@processingQueueVariables
-    else c("msLevel", "centroided")
-}
-
-#' Updated version that supports arbitrary spectra variables.
-#'
-#' @param spectraVariables `character` defining spectra variables that should
-#'     be passed to the function.
-#'
-#' @noRd
-.peaksapply2 <- function(object, FUN = NULL, ...,
-                         spectraVariables = .processingQueueVariables(object),
-                         f = dataStorage(object), BPPARAM = bpparam()) {
+.peaksapply <- function(object, FUN = NULL, ...,
+                        spectraVariables = .processingQueueVariables(object),
+                        f = dataStorage(object), BPPARAM = bpparam()) {
     len <- length(object)
     if (!len)
         return(list())
@@ -168,7 +162,7 @@ addProcessing <- function(object, FUN, ..., spectraVariables = character()) {
             if (length(svars))
                 spd <- as.data.frame(spectraData(z, columns = svars))
             else spd <- NULL
-            .apply_processing_queue2(peaksData(z), spd, queue = queue)
+            .apply_processing_queue(peaksData(z), spd, queue = queue)
         }, queue = pqueue, svars = spectraVariables, BPPARAM = BPPARAM)
         unsplit(res, f = f, drop = TRUE)
     } else peaksData(object@backend)
@@ -230,11 +224,14 @@ applyProcessing <- function(object, f = dataStorage(object),
     if (length(f) != length(object))
         stop("length 'f' has to be equal to the length of 'object' (",
              length(object), ")")
-    bknds <- bplapply(split(object@backend, f = f), function(z, queue) {
-        peaksData(z) <- .apply_processing_queue(peaksData(z), msLevel(z),
-                                                centroided(z), queue)
+    bknds <- bplapply(split(object@backend, f = f), function(z, queue, svars) {
+        if (length(svars))
+            spd <- as.data.frame(spectraData(z, columns = svars))
+        else spd <- NULL
+        peaksData(z) <- .apply_processing_queue(peaksData(z), spd, queue)
         z
-    }, queue = object@processingQueue, BPPARAM = BPPARAM)
+    }, queue = object@processingQueue,
+    svars = .processingQueueVariables(object), BPPARAM = BPPARAM)
     bknds <- backendMerge(bknds)
     if (is.unsorted(f))
         bknds <- bknds[order(unlist(split(seq_along(bknds), f),